+++ spinn_front_end_common/interface/buffer_management/__init__.py
@@ -1,6 +1,4 @@
 
-from .abstract_database import AbstractDatabase
 from .buffer_manager import BufferManager
-from .sqllite_database import SqlLiteDatabase
 
-__all__ = ["AbstractDatabase", "BufferManager", "SqlLiteDatabase"]
+__all__ = ["BufferManager"]
