+++ spinn_front_end_common/interface/buffer_management/storage_objects/abstract_database.py
